Form now works as expected.

+++ src/core/homepage_elements/featured/views.py
@@ -1,6 +1,7 @@
 from django.urls import reverse
 from django.http import HttpResponse
 from django.shortcuts import redirect, render, get_object_or_404
+from django.contrib import messages
 
 from core.homepage_elements.featured import models, forms
 from submission import models as submission_models
@@ -15,27 +16,41 @@ def featured_articles(request):
     featured_article_pks = [f.article.pk for f in featured_arts.all()]
     articles = submission_models.Article.objects.filter(date_published__isnull=False,
                                                         journal=request.journal).exclude(pk__in=featured_article_pks)
-    form = forms.FeaturedForm()
+    form = forms.FeaturedForm(journal=request.journal)
 
-    if 'article_id' in request.POST:
-        article_id = request.POST.get('article_id')
-        article = get_object_or_404(submission_models.Article, pk=article_id, journal=request.journal)
+    if request.POST:
 
-        models.FeaturedArticle.objects.create(
-            article=article,
-            journal=request.journal,
-            added_by=request.user,
-            sequence=request.journal.next_featured_article_order()
-        )
+        if 'article_id' in request.POST:
+            article_id = request.POST.get('article_id')
+            article = get_object_or_404(submission_models.Article, pk=article_id, journal=request.journal)
 
-        return redirect(reverse('featured_articles_setup'))
+            models.FeaturedArticle.objects.create(
+                article=article,
+                journal=request.journal,
+                added_by=request.user,
+                sequence=request.journal.next_featured_article_order()
+            )
 
-    if 'delete' in request.POST:
-        article_id = request.POST.get('delete')
-        featured_article = get_object_or_404(models.FeaturedArticle, article__pk=article_id,
-                                             journal=request.journal)
-        featured_article.delete()
-        return redirect(reverse('featured_articles_setup'))
+            return redirect(reverse('featured_articles_setup'))
+
+        if 'delete' in request.POST:
+            article_id = request.POST.get('delete')
+            featured_article = get_object_or_404(models.FeaturedArticle, article__pk=article_id,
+                                                 journal=request.journal)
+            featured_article.delete()
+            return redirect(reverse('featured_articles_setup'))
+
+        if 'form' in request.POST:
+            form = forms.FeaturedForm(request.POST, journal=request.journal)
+
+            if form.is_valid():
+                form.save(journal=request.journal)
+                messages.add_message(
+                    request,
+                    messages.SUCCESS,
+                    'Settings saved.',
+                )
+                return redirect(reverse('featured_articles_setup'))
 
     template = 'featured_setup.html'
     context = {
